Plump test fixture variables to outputs
The `terraform` verifier will transfer Terraform outputs into Inspec as
attributes. However, we were depending on the value of those variables
to run tests against the Terraform instantiated resources, and didn't
have access to that data.

This commit solves the issue by passing all Terraform variables back out
as outputs, so that the InSpec tests can easily reference that data.